Renamed BanTask to BanExpirationTask for clarity (Fixes #32)
